+++ source-code/app/main.js
@@ -1,11 +1,11 @@
+import { enableProdMode } from '@angular/core';
 // The browser platform with a compiler
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from "./app.module";
-import { enableDebugTools } from '@angular/platform-browser';
-// enableProdMode();
+enableProdMode();
 // Compile and launch the module
-platformBrowserDynamic().bootstrapModule(AppModule)
-    .then(function (appRef) {
-    enableDebugTools(appRef);
-});
+platformBrowserDynamic().bootstrapModule(AppModule);
+// .then((appRef) => {
+//     enableDebugTools(appRef);
+// }); 
 //# sourceMappingURL=main.js.map