+++ source-code/app/main.ts
@@ -3,9 +3,9 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import {AppModule} from "./app.module";
 import {enableDebugTools} from '@angular/platform-browser';
-// enableProdMode();
+enableProdMode();
 // Compile and launch the module
-platformBrowserDynamic().bootstrapModule(AppModule)
-    .then((appRef) => {
-        enableDebugTools(appRef);
-    });
+platformBrowserDynamic().bootstrapModule(AppModule);
+    // .then((appRef) => {
+    //     enableDebugTools(appRef);
+    // });